Drop empty `host`
Looks like this might not be needed as we are getting separate build and
host environments anyways (please see ref below). So go ahead and drop
the empty `host` to make things read a little nicer. Also fixes the
linter error (though it appears the linter may have a bug).

ref: https://circleci.com/api/v1.1/project/github/conda-forge/binutils-feedstock/24/output/104/0?file=true